创建客户相关类

+++ ErpSystem/src/com/erpsystem/service/IOrderService.java
@@ -33,7 +33,7 @@ public interface IOrderService {
 	 * @return
 	 * @throws SQLException
 	 */
-	public Order findById(Integer id) throws SQLException;
+	public Order findById(Long id) throws SQLException;
 	
 	/**
 	 * 插入一条订单
